fix(components): adjust validation of state to happen inside parent with force update, adjust e2e and unit tests | as / bh / mh | #1861
To properly batch updates e.g. in frameworks that use .map or ngFor or v-for and the likes, it is important to not immediately do the validation of the state, because it would cause to iteratively throw errors, while the rerender happens. Thus force update will now be invoked from the child, which will do the validation once all the children states are updated.